Renaming client.py to bot.py
